requirements.txt
absl-py==0.13.0
addict==2.4.0
AdelaiDet==0.2.0
albumentations==1.0.3
alembic==1.6.5
altgraph==0.17
antlr4-python3-runtime==4.8
appdirs==1.4.4
astor==0.8.1
asynctest==0.13.0
attr==0.3.1
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.3
cityscapesScripts==2.2.0
click==8.0.1
cliff==3.8.0
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.1.2
codecov==2.1.11
colorama==0.4.4
coloredlogs==15.0.1
colorlog==5.0.1
contextlib2==21.6.0
coremltools==4.0
coverage==5.5
cycler==0.10.0
Cython==0.29.24
dataclasses==0.8
decorator==4.4.2
detectron2==0.4
filelock==3.0.12
flake8==3.9.2
func-timeout==4.3.5
future==0.18.2
fvcore==0.1.5.post20211019
google-auth==1.34.0
google-auth-oauthlib==0.4.5
greenlet==1.1.0
grpcio==1.39.0
h5py==3.1.0
humanfriendly==9.2
hyperopt==0.1.2
idna==3.2
imagecorruptions==1.1.2
imageio==2.9.0
imgaug==0.4.0
imgviz==1.4.0
importlib-metadata==4.6.1
iniconfig==1.1.1
interrogate==1.4.0
iopath==0.1.9
isort==4.3.21
joblib==1.0.1
json-tricks==3.15.5
kiwisolver==1.3.1
kwarray==0.5.19
labelme==4.5.13
leveldb==0.201
lmdb==1.2.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.2
mccabe==0.6.1
mmcv-full==1.0.5
mmlvis==10.5.3
mmpycocotools==12.0.3
mpmath==1.2.1
netifaces==0.11.0
networkx==2.5.1
nni==2.3
numpy==1.19.5
nvidia-ml-py==11.460.79
oauthlib==3.1.1
omegaconf==2.1.1
onnx==1.6.0
onnxruntime==1.6.0
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
optuna==2.8.0
ordered-set==4.0.2
packaging==21.0
pandas==1.1.5
pbr==5.6.0
Pillow==8.3.1
pluggy==0.13.1
Polygon3==3.0.9.1
portalocker==2.3.2
prettytable==2.1.0
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydot==1.4.2
pyflakes==2.3.1
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pyiou==0.0.2
pymongo==3.12.0
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.9
pytest==6.2.4
python-dateutil==2.8.2
python-editor==1.0.4
PythonWebHDFS==0.2.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
QtPy==1.11.2
rapidfuzz==1.8.0
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.13.3
rsa==4.7.2
schema==0.7.4
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.7.1
simplejson==3.17.3
six==1.16.0
SQLAlchemy==1.4.21
stevedore==3.3.0
sympy==1.8
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminaltables==3.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2020.9.3
toml==0.10.2
torch @ file:///data/xingyueqi/bak_package/torch-1.7.1%2Bcu101-cp36-cp36m-linux_x86_64.whl
torchaudio @ file:///data/xingyueqi/bak_package/torchaudio-0.7.2-cp36-cp36m-linux_x86_64.whl
torchvision @ file:///data/xingyueqi/bak_package/torchvision-0.8.2%2Bcu101-cp36-cp36m-linux_x86_64.whl
tqdm==4.61.2
typing==3.7.4.3
typing-extensions==3.10.0.0
ubelt==0.9.5
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==0.58.0
websockets==9.1
Werkzeug==2.0.1
xdoctest==0.15.5
yacs==0.1.8
yapf==0.31.0
zipp==3.5.0
